+++ src/pages/index.astro
@@ -1,10 +1,10 @@
 ---
 import AstroLayout from "../layouts/layout.astro";
 import Home from "../components/Home/Home";
-import { getAndParseHyAllPosts } from "../utils/hygraph";
+import { getAndParseHyAllPosts } from "../services/hygraph";
 import { Effect } from "effect";
-import { getAndParseRecentTraks } from "../utils/audioscrobbler";
-import { program } from "../utils/letterboxd";
+import { getAndParseRecentTraks } from "../services/audioscrobbler";
+import { program } from "../services/letterboxd";
 
 const music = await Effect.runPromise(getAndParseRecentTraks())
   .then((data) => data)
