+++ src/services/letterboxd.ts
@@ -1,10 +1,28 @@
 import * as S from "@effect/schema/Schema";
 import { parseString } from "xml2js";
 import { Effect, pipe } from "effect";
-import { getText } from "./utils";
-import { decodeUnknownEither } from "./decode";
-import { RssSchema } from "../models/model";
+import { getText } from "../utils/utils";
+import { decodeUnknownEither } from "../utils/decode";
 import { getAndParseMovieById } from "./tmdb";
+
+
+export const RssSchema = S.struct({
+    rss: S.struct({
+        channel: S.array(
+            S.struct({
+                item: S.array(
+                    S.struct({
+                        title: S.array(S.string),
+                        link: S.array(S.string),
+                        description: S.array(S.string),
+                        "tmdb:movieId": S.optional(S.array(S.string)),
+                    })
+                )
+            }))
+    })
+})
+
+export type TRss = S.Schema.Type<typeof RssSchema>
 
 const parseXmlContent = (xmlContent: string): Promise<unknown> =>
     new Promise((resolve, reject) => {
